Working parser
Little changes in the searched strings, working properly on the testing set. Don't forget to define the references array (first line of the code) and call the revnum function first, it's output is then the second parameter of the parser function.

Parsing of the parameter prov:wasDerivedFrom doesn't work correctly, but the others are working properly.